docs: replace vuelidate references with regle
