+++ packages/docs/package.json
@@ -22,10 +22,10 @@
   },
   "dependencies": {
     "@cosmicjs/sdk": "^1.5.2",
+    "@regle/core": "1.8.4",
+    "@regle/rules": "1.8.4",
     "@sentry/vue": "^9.12.0",
     "@vue/compiler-dom": "^3.5.13",
-    "@vuelidate/core": "^2.0.3",
-    "@vuelidate/validators": "^2.0.4",
     "@vuetify/one": "^2.3.0",
     "algoliasearch": "^4.24.0",
     "fflate": "^0.8.2",
